ui i18n; listing alerts

+++ src/main/scala/bg/statealerts/services/AlertService.scala
@@ -31,7 +31,7 @@ class AlertService {
   }
 
   @Transactional
-  def delete(id: Long) = {
+  def delete(id: Int) = {
     dao.delete(classOf[Alert], id)
   }
 }
